refactor : 베팅 관련 로직에서 Dealer, Player를 추상화해서 사용
