+++ src/main/java/model/participant/role/Gameable.java
@@ -1,6 +1,7 @@
-package model.deck;
+package model.participant.role;
 
 import java.util.List;
+import model.deck.Card;
 
 public interface Gameable {
     void receiveCard(final Card card);
